unit_test: add eBPF tests
For every C file in the tests/ebpf directory, we compile it to eBPF with
clang and send it to the device. Then we write the input and check if
the memory matches what is expected. Programs are expected to always
return 0.

Both the input and expected memory are optional, but if used they must
have the same filename as the test but with extensions ".in" and ".mem"
respectively.

One limitation of the current approach is that we can't run the same
program against multiple inputs, but we plan to address this in the
future.

As an illustration, we add a simple test case that just increments every
element of an array of integers.