driver: add fsync support
The fsync syscall can be used to trigger program execution on a Hermes
device. The program and data slots must have been previously populated.

We use a waitqueue per eBPF engine to wait for program completion. When
an engine finishes execution, the hardware must:
1. Set the EHCMDDONE register for that engine
2. Send an interrupt using the IRQ vector number for that engine.

When the interrupt is received, the wait queue is woken up and checks
the EHCMDDONE register.

Note that a bugged or malicious program that never finishes will never
cause the wake up. This can be addressed in the future by using
wake_up_interruptible_timeout, but it will need hardware support to halt
the running program (which we don't have yet).

Initially we always execute on eBPF engine 0, but this will be addressed
on another commit.

A word on the eBPF return code: we expect programs to exit with return
code 0: anything else is considered an error and the fsync returns with
-ENOEXEC. If the return code is needed, then the eBPF program should put
it somewhere in the data slot and then read() it back.