driver: add eBPF interrupts
The eBPF engines will raise an interrupt when a program has exited.

This patch initially registers one interrupt per eBPF engine, with an
empty handler, which will be filled in future patches.

The pointer argument passed to the IRQ handler contains the
hermes_dev and the engine ID related to that interrupt as those will be
needed to handle the interrupt in future patches.